Books validations

+++ Books.js
@@ -2,7 +2,7 @@
  * Kaip rasyti JSDOc'sus?
  * Link: https://jsdoc.app
  */
-
+const Validation = require('./Validations');
 const Books = {};
 
 /**
@@ -14,7 +14,18 @@ const Books = {};
  * @returns {Promise<string>} Tekstas, apibudinantis, koks autorius ir kurias metais isleido knyga.
  */
 Books.create = async (connection, authorId, bookName, bookReleaseYear) => {
-    sql = 'INSERT INTO books (authorId,title, releaseYear)\
+    //VALIDATIONS
+    if (!Validation.IDisValid(authorId)) {
+        return `Autoriaus ID turi buti teigiamas sveikasis skaicius!`;
+    }
+    if (!Validation.isText(bookName)) {
+        return `Parametras turi buti ne tuscias tekstas!`;
+    }
+    if (!Validation.isYearValid(bookReleaseYear)) {
+        return `Parametras turi buti teigiamas sveikasis skaicius!`;
+    }
+
+    const sql = 'INSERT INTO books (authorId,title, releaseYear)\
     VALUES ("'+ authorId + '", "' + bookName + '", "' + bookReleaseYear + '")';
     [rows] = await connection.execute(sql);
     const createBook = `Was added to the book list: author with ID ${authorId}, book title "${bookName}", release year ${bookReleaseYear}.`
@@ -27,8 +38,8 @@ Books.create = async (connection, authorId, bookName, bookReleaseYear) => {
  * @returns {Promise<Object[]>} Tekstas, apibudinantis, koks autorius ir kurias metais isleido knyga.
  */
 Books.listAll = async (connection) => {
-    sql = 'SELECT * FROM books';
-    [rows] = await connection.execute(sql);
+    const sql = 'SELECT * FROM books';
+    const [rows] = await connection.execute(sql);
     const booksList = [];
     let i = 0;
     for (const book of rows) {
@@ -45,8 +56,13 @@ Books.listAll = async (connection) => {
  * @returns {Promise<string>} Tekstas su knygos duomenimis..
  */
 Books.findByName = async (connection, bookName) => {
-    sql = 'SELECT * FROM `books` WHERE `title` = "' + bookName + '"';
-    [rows] = await connection.execute(sql);
+    //VALIDATIONS
+    if (!Validation.isText(bookName)) {
+        return `Parametras turi buti ne tuscias tekstas!`;
+    }
+
+    const sql = 'SELECT * FROM `books` WHERE `title` = "' + bookName + '"';
+    const [rows] = await connection.execute(sql);
     const byTitle = `Book "${bookName}" is written by author_Id ${rows[0].authorId}, released in year ${rows[0].releaseYear}.`;
     return byTitle;
 }
@@ -58,8 +74,13 @@ Books.findByName = async (connection, bookName) => {
  * @returns {Promise<string>} Tekstas su knygos duomenimis..
  */
 Books.findByAuthorId = async (connection, authorId) => {
-    sql = 'SELECT * FROM `books` WHERE `authorId` =' + authorId;
-    [rows] = await connection.execute(sql);
+    //VALIDATIONS
+    if (!Validation.IDisValid(authorId)) {
+        return `Autoriaus ID turi buti teigiamas sveikasis skaicius!`;
+    }
+
+    const sql = 'SELECT * FROM `books` WHERE `authorId` =' + authorId;
+    const [rows] = await connection.execute(sql);
 
     const booksListByAuthorId = [];
     let i = 0;
@@ -76,9 +97,14 @@ Books.findByAuthorId = async (connection, authorId) => {
  * @returns {Promise<Object[]>} Sarasas su knygu objektais.
  */
 Books.findByYear = async (connection, bookReleaseYear) => {
-    sql = 'SELECT * FROM `books` WHERE `releaseYear` = "' + bookReleaseYear + '"\
+    //VALIDATIONS
+    if (!Validation.isYearValid(bookReleaseYear)) {
+        return `Parametras turi buti teigiamas sveikasis skaicius!`;
+    }
+
+    const sql = 'SELECT * FROM `books` WHERE `releaseYear` = "' + bookReleaseYear + '"\
     ORDER BY `authorId` ASC';
-    [rows] = await connection.execute(sql);
+    const [rows] = await connection.execute(sql);
 
     const booksListByYear = [];
     let i = 0;
@@ -96,7 +122,18 @@ Books.findByYear = async (connection, bookReleaseYear) => {
  * @returns {Promise<Object[]>} Sarasas su knygu objektais.
  */
 Books.updateById = async (connection, bookId, propertyName, propertyValue) => {
-    sql = 'UPDATE books SET ' + propertyName + ' = "' + propertyValue + '" WHERE books.id =' + bookId;
+    //VALIDATIONS
+    if (!Validation.IDisValid(bookId)) {
+        return `Knygos ID turi buti teigiamas sveikasis skaicius!`;
+    }
+    if (!Validation.isText(propertyName)) {
+        return `Parametras turi buti ne tuscias tekstas!`;
+    }
+    if (!Validation.isText(propertyValue)) {
+        return `Parametras turi buti ne tuscias tekstas!`;
+    }
+
+    const sql = 'UPDATE books SET ' + propertyName + ' = "' + propertyValue + '" WHERE books.id =' + bookId;
     [rows] = await connection.execute(sql);
     const updatedBookByID = `Book with ID ${bookId} got ${propertyName} changed to "${propertyValue}."`
     return updatedBookByID;
@@ -106,10 +143,18 @@ Books.updateById = async (connection, bookId, propertyName, propertyValue) => {
  * @param {Object} connection Objektas, su kuriuo kvieciame duombazes mainpuliavimo metodus.
  * @param {number} bookId Knygos ID.
  * @param {string} bookName Atnaujinamos savybes pavadinimas.
- *  @returns {Promise<string>} Tekstas su knygos duomenimis.
+ * @returns {Promise<string>} Tekstas su knygos duomenimis.
  */
 Books.updateNameById = async (connection, bookId, bookName) => {
-    sql = 'UPDATE books SET title = "' + bookName + '" WHERE books.id =' + bookId;
+    //VALIDATIONS
+    if (!Validation.IDisValid(bookId)) {
+        return `Knygos ID turi buti teigiamas sveikasis skaicius!`;
+    }
+    if (!Validation.isText(bookName)) {
+        return `Parametras turi buti ne tuscias tekstas!`;
+    }
+
+    const sql = 'UPDATE books SET title = "' + bookName + '" WHERE books.id =' + bookId;
     [rows] = await connection.execute(sql);
     const updatedTitleById = `Book with ID ${bookId} has a new title now as "${bookName}."`
     return updatedTitleById;
@@ -119,10 +164,18 @@ Books.updateNameById = async (connection, bookId, bookName) => {
  * @param {Object} connection Objektas, su kuriuo kvieciame duombazes mainpuliavimo metodus.
  * @param {number} bookId Knygos ID.
  * @param {number} bookReleaseYear Knygos isleidimo metai.
- *  @returns {Promise<string>} Tekstas su knygos duomenimis.
+ * @returns {Promise<string>} Tekstas su knygos duomenimis.
  */
 Books.updateYearById = async (connection, bookId, bookReleaseYear) => {
-    sql = 'UPDATE books SET releaseYear = "' + bookReleaseYear + '" WHERE books.id =' + bookId;
+    //VALIDATIONS
+    if (!Validation.IDisValid(bookId)) {
+        return `Knygos ID turi buti teigiamas sveikasis skaicius!`;
+    }
+    if (!Validation.isYearValid(bookReleaseYear)) {
+        return `Parametras turi buti teigiamas sveikasis skaicius!`;
+    }
+
+    const sql = 'UPDATE books SET releaseYear = "' + bookReleaseYear + '" WHERE books.id =' + bookId;
     [rows] = await connection.execute(sql);
     const updatedBookYearByID = `Book with ID ${bookId} has a new release year ${bookReleaseYear} now."`
     return updatedBookYearByID;
@@ -131,10 +184,15 @@ Books.updateYearById = async (connection, bookId, bookReleaseYear) => {
  * Knygos istrynimas pagal ID.
  * @param {Object} connection Objektas, su kuriuo kvieciame duombazes mainpuliavimo metodus.
  * @param {number} bookId Knygos ID.
- *  @returns {Promise<string>} Tekstas su knygos duomenimis.
+ * @returns {Promise<string>} Tekstas su knygos duomenimis.
  */
 Books.delete = async (connection, bookId) => {
-    sql = 'DELETE FROM books WHERE books.id =' + bookId;
+    //VALIDATIONS
+    if (!Validation.IDisValid(bookId)) {
+        return `Knygos ID turi buti teigiamas sveikasis skaicius!`;
+    }
+
+    const sql = 'DELETE FROM books WHERE books.id =' + bookId;
     [rows] = await connection.execute(sql);
     const deletedBook = `Book with ID ${bookId} has been removed from books list!`
     return deletedBook;
@@ -143,10 +201,15 @@ Books.delete = async (connection, bookId) => {
  * Visu knygu istrynimas pagal autoriaus ID.
  * @param {Object} connection Objektas, su kuriuo kvieciame duombazes mainpuliavimo metodus.
  * @param {number} authorId Autoriaus ID.
- *  @returns {Promise<string>} Tekstas su knygos duomenimis.
+ * @returns {Promise<string>} Tekstas su knygos duomenimis.
  */
 Books.deleteAllByAuthorId = async (connection, authorId) => {
-    sql = 'DELETE FROM books WHERE authorId =' + authorId;
+    //VALIDATIONS
+    if (!Validation.IDisValid(authorId)) {
+        return `Knygos ID turi buti teigiamas sveikasis skaicius!`;
+    }
+
+    const sql = 'DELETE FROM books WHERE authorId =' + authorId;
     [rows] = await connection.execute(sql);
     const deletedBookByAuthorId = `All books of author ID ${authorId} have been deleted from books list.`
     return deletedBookByAuthorId;
